Make project a multi-app monolith
