+++ cmd/devops/cmd/cors_check.go
