+++ cmd/kubectl-plugins/kubectl-list_addresses/main.go
@@ -17,11 +17,12 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package kubectl_lookup_address
+package main
 
 import (
+	"fmt"
 	"log/slog"
-	"slices"
+	"os"
 
 	"github.com/ginolatorilla/devops/pkg/kubectlplugin"
 	"github.com/spf13/cobra"
@@ -32,67 +33,90 @@ import (
 	gocoreV1 "k8s.io/client-go/kubernetes/typed/core/v1"
 )
 
-func NewCommand(runnerOpts ...kubectlplugin.RunnerOpts) *cobra.Command {
+func main() {
+	if err := newCommand(kubectlplugin.WithDefaultKubeApi()).Execute(); err != nil {
+		fmt.Fprintf(os.Stderr, "error: %v\n", err)
+		os.Exit(1)
+	}
+}
+
+func newCommand(runnerOpts ...kubectlplugin.RunnerOpts) *cobra.Command {
 	return kubectlplugin.
-		NewRunner(lookupAddress, append(runnerOpts, kubectlplugin.WithResourcePrinters())...).
+		NewRunner(listAddresses, append(runnerOpts, kubectlplugin.WithResourcePrinters())...).
 		ToCobraCommand(
-			"kubectl-lookup_address",
-			"Finds Kubernetes resources by IP address",
-			kubectlplugin.WithArgs(cobra.ExactArgs(1), []string{"ip-address"}),
+			"kubectl-list_addresses",
+			"Lists all IP addresses in the cluster",
 		)
 }
 
-func lookupAddress(a kubectlplugin.HandlerArgs) (runtime.Object, error) {
-	ipAddress := a.Args[0]
+func listAddresses(a kubectlplugin.HandlerArgs) (runtime.Object, error) {
 	client := a.KubeApi.CoreV1()
 
-	tableBuilder := kubectlplugin.
-		NewTableBuilder().
-		AdditionalColumns(
-			kubectlplugin.ResourceKindColumn,
-			kubectlplugin.Column{Name: "Type", Description: "The type of address"},
-		)
+	tableBuilder := kubectlplugin.NewTableBuilder().AdditionalColumns(
+		kubectlplugin.ResourceKindColumn,
+		kubectlplugin.Column{Name: "Type", Description: "The type of address"},
+		kubectlplugin.Column{Name: "Address", Description: "The IP address"},
+	)
 
-	if err := recordServicesWithMatchingAddress(a, client, ipAddress, tableBuilder); err != nil {
+	if err := getServiceAddresses(a, client, tableBuilder); err != nil {
 		slog.Warn("failed to list services", "error", err)
 	}
-	if err := recordPodsWithMatchingAddress(a, client, ipAddress, tableBuilder); err != nil {
+
+	if err := getPodAddresses(a, client, tableBuilder); err != nil {
 		slog.Warn("failed to list pods", "error", err)
 	}
-	if err := recordNodesWithMatchingAddress(a, client, ipAddress, tableBuilder); err != nil {
+
+	if err := getNodeAddresses(a, client, tableBuilder); err != nil {
 		slog.Warn("failed to list nodes", "error", err)
 	}
+
 	return &tableBuilder.Table, nil
 }
 
-func recordServicesWithMatchingAddress(
+func getServiceAddresses(
 	a kubectlplugin.HandlerArgs,
 	client gocoreV1.CoreV1Interface,
-	ipAddress string,
 	tableBuilder *kubectlplugin.TableBuilder,
 ) error {
 	services, err := client.Services(a.Namespace).List(a.Cmd.Context(), metaV1.ListOptions{})
 	if err != nil {
 		return err
 	}
+
 	for _, service := range services.Items {
-		if slices.Contains(service.Spec.ClusterIPs, ipAddress) {
-			tableBuilder.AddRow(&service, map[string]any{"Type": "ClusterIP", "Kind": "Service"})
+		for _, ip := range service.Spec.ClusterIPs {
+			if ip == "" || ip == "None" {
+				continue
+			}
+			tableBuilder.AddRow(&service, map[string]any{
+				"Kind":    "Service",
+				"Type":    "ClusterIP",
+				"Address": ip,
+			})
 		}
-		if slices.Contains(service.Spec.ExternalIPs, ipAddress) {
-			tableBuilder.AddRow(&service, map[string]any{"Type": "ExternalIP", "Kind": "Service"})
+
+		for _, ip := range service.Spec.ExternalIPs {
+			tableBuilder.AddRow(&service, map[string]any{
+				"Kind":    "Service",
+				"Type":    "ExternalIP",
+				"Address": ip,
+			})
 		}
-		if service.Spec.LoadBalancerIP == ipAddress {
-			tableBuilder.AddRow(&service, map[string]any{"Type": "LoadBalancerIP", "Kind": "Service"})
+
+		if service.Spec.LoadBalancerIP != "" {
+			tableBuilder.AddRow(&service, map[string]any{
+				"Kind":    "Service",
+				"Type":    "LoadBalancerIP",
+				"Address": service.Spec.LoadBalancerIP,
+			})
 		}
 	}
 	return nil
 }
 
-func recordPodsWithMatchingAddress(
+func getPodAddresses(
 	a kubectlplugin.HandlerArgs,
 	client gocoreV1.CoreV1Interface,
-	ipAddress string,
 	tableBuilder *kubectlplugin.TableBuilder,
 ) error {
 	pods, err := client.Pods(a.Namespace).List(a.Cmd.Context(), metaV1.ListOptions{})
@@ -100,19 +124,20 @@ func recordPodsWithMatchingAddress(
 		return err
 	}
 	for _, pod := range pods.Items {
-		if slices.ContainsFunc(pod.Status.PodIPs, func(p coreV1.PodIP) bool {
-			return p.IP == ipAddress
-		}) {
-			tableBuilder.AddRow(&pod, map[string]any{"Type": "PodIP", "Kind": "Pod"})
+		for _, ip := range pod.Status.PodIPs {
+			tableBuilder.AddRow(&pod, map[string]any{
+				"Kind":    "Pod",
+				"Type":    "PodIP",
+				"Address": ip.IP,
+			})
 		}
 	}
 	return nil
 }
 
-func recordNodesWithMatchingAddress(
+func getNodeAddresses(
 	a kubectlplugin.HandlerArgs,
 	client gocoreV1.CoreV1Interface,
-	ipAddress string,
 	tableBuilder *kubectlplugin.TableBuilder,
 ) error {
 	nodes, err := client.Nodes().List(a.Cmd.Context(), metaV1.ListOptions{})
@@ -120,10 +145,15 @@ func recordNodesWithMatchingAddress(
 		return err
 	}
 	for _, node := range nodes.Items {
-		for _, n := range node.Status.Addresses {
-			if n.Address == ipAddress {
-				tableBuilder.AddRow(&node, map[string]any{"Type": n.Type, "Kind": "Node"})
+		for _, address := range node.Status.Addresses {
+			if address.Type != coreV1.NodeInternalIP && address.Type != coreV1.NodeExternalIP {
+				continue
 			}
+			tableBuilder.AddRow(&node, map[string]any{
+				"Kind":    "Node",
+				"Type":    address.Type,
+				"Address": address.Address,
+			})
 		}
 	}
 	return nil
