+++ cmd/kubectl-plugins/kubectl-list_certs/main.go
@@ -17,12 +17,13 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package kubectl_list_certs
+package main
 
 import (
 	"crypto/tls"
 	"fmt"
 	"log/slog"
+	"os"
 
 	"github.com/dustin/go-humanize"
 	"github.com/ginolatorilla/devops/pkg/kubectlplugin"
@@ -32,7 +33,14 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 )
 
-func NewCommand(runnerOpts ...kubectlplugin.RunnerOpts) *cobra.Command {
+func main() {
+	if err := newCommand(kubectlplugin.WithDefaultKubeApi()).Execute(); err != nil {
+		fmt.Fprintf(os.Stderr, "error: %v\n", err)
+		os.Exit(1)
+	}
+}
+
+func newCommand(runnerOpts ...kubectlplugin.RunnerOpts) *cobra.Command {
 	return kubectlplugin.
 		NewRunner(listCerts, append(runnerOpts, kubectlplugin.WithResourcePrinters())...).
 		ToCobraCommand(
