+++ cmd/kubectl-plugins/kubectl-list_certs/testdata/invalid/kubeConfig.yaml
