+++ cmd/kubectl-plugins/kubectl-list_certs/testdata/missing-namespace/kubeConfig.yaml
