+++ cmd/kubectl-plugins/kubectl-list_certs/testdata/ok/kubeConfig.yaml
