+++ cmd/kubectl-plugins/kubectl-list_certs/testdata/ok/tls.crt
