+++ cmd/kubectl-plugins/kubectl-list_unhealthy_pods/main.go
@@ -17,10 +17,11 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package kubectl_list_unhealthy_pods
+package main
 
 import (
 	"fmt"
+	"os"
 	"slices"
 
 	"github.com/ginolatorilla/devops/pkg/kubectlplugin"
@@ -31,7 +32,14 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 )
 
-func NewCommand(runnerOpts ...kubectlplugin.RunnerOpts) *cobra.Command {
+func main() {
+	if err := newCommand(kubectlplugin.WithDefaultKubeApi()).Execute(); err != nil {
+		fmt.Fprintf(os.Stderr, "error: %v\n", err)
+		os.Exit(1)
+	}
+}
+
+func newCommand(runnerOpts ...kubectlplugin.RunnerOpts) *cobra.Command {
 	return kubectlplugin.
 		NewRunner(listUnhealthyPods, append(runnerOpts, kubectlplugin.WithResourcePrinters())...).
 		ToCobraCommand(
