+++ cmd/kubectl-plugins/kubectl-list_unhealthy_pods/main_test.go
@@ -1,4 +1,4 @@
-package kubectl_list_unhealthy_pods
+package main
 
 import (
 	"context"
@@ -22,7 +22,7 @@ func TestNewCommand(t *testing.T) {
 		client := fake.NewClientset()
 		loadResources(t, client, "test")
 
-		cmd := NewCommand(kubectlplugin.WithKubeApi(client))
+		cmd := newCommand(kubectlplugin.WithKubeApi(client))
 		cmd.SetArgs([]string{"-n", "test"})
 		assert.NoError(cmd.Execute())
 	})
@@ -32,7 +32,7 @@ func TestNewCommand(t *testing.T) {
 		loadCannedError(t, client, "list", "pods")
 		loadResources(t, client, "test")
 
-		cmd := NewCommand(kubectlplugin.WithKubeApi(client))
+		cmd := newCommand(kubectlplugin.WithKubeApi(client))
 		cmd.SetArgs([]string{"-n", "test"})
 		assert.Error(cmd.Execute())
 	})
