+++ cmd/kubectl-plugins/kubectl-lookup_address/main.go
@@ -17,10 +17,13 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package kubectl_list_addresses
+package main
 
 import (
+	"fmt"
 	"log/slog"
+	"os"
+	"slices"
 
 	"github.com/ginolatorilla/devops/pkg/kubectlplugin"
 	"github.com/spf13/cobra"
@@ -31,83 +34,74 @@ import (
 	gocoreV1 "k8s.io/client-go/kubernetes/typed/core/v1"
 )
 
-func NewCommand(runnerOpts ...kubectlplugin.RunnerOpts) *cobra.Command {
+func main() {
+	if err := newCommand(kubectlplugin.WithDefaultKubeApi()).Execute(); err != nil {
+		fmt.Fprintf(os.Stderr, "error: %v\n", err)
+		os.Exit(1)
+	}
+}
+
+func newCommand(runnerOpts ...kubectlplugin.RunnerOpts) *cobra.Command {
 	return kubectlplugin.
-		NewRunner(listAddresses, append(runnerOpts, kubectlplugin.WithResourcePrinters())...).
+		NewRunner(lookupAddress, append(runnerOpts, kubectlplugin.WithResourcePrinters())...).
 		ToCobraCommand(
-			"kubectl-list_addresses",
-			"Lists all IP addresses in the cluster",
+			"kubectl-lookup_address",
+			"Finds Kubernetes resources by IP address",
+			kubectlplugin.WithArgs(cobra.ExactArgs(1), []string{"ip-address"}),
 		)
 }
 
-func listAddresses(a kubectlplugin.HandlerArgs) (runtime.Object, error) {
+func lookupAddress(a kubectlplugin.HandlerArgs) (runtime.Object, error) {
+	ipAddress := a.Args[0]
 	client := a.KubeApi.CoreV1()
 
-	tableBuilder := kubectlplugin.NewTableBuilder().AdditionalColumns(
-		kubectlplugin.ResourceKindColumn,
-		kubectlplugin.Column{Name: "Type", Description: "The type of address"},
-		kubectlplugin.Column{Name: "Address", Description: "The IP address"},
-	)
+	tableBuilder := kubectlplugin.
+		NewTableBuilder().
+		AdditionalColumns(
+			kubectlplugin.ResourceKindColumn,
+			kubectlplugin.Column{Name: "Type", Description: "The type of address"},
+		)
 
-	if err := getServiceAddresses(a, client, tableBuilder); err != nil {
+	if err := recordServicesWithMatchingAddress(a, client, ipAddress, tableBuilder); err != nil {
 		slog.Warn("failed to list services", "error", err)
 	}
-
-	if err := getPodAddresses(a, client, tableBuilder); err != nil {
+	if err := recordPodsWithMatchingAddress(a, client, ipAddress, tableBuilder); err != nil {
 		slog.Warn("failed to list pods", "error", err)
 	}
-
-	if err := getNodeAddresses(a, client, tableBuilder); err != nil {
+	if err := recordNodesWithMatchingAddress(a, client, ipAddress, tableBuilder); err != nil {
 		slog.Warn("failed to list nodes", "error", err)
 	}
-
 	return &tableBuilder.Table, nil
 }
 
-func getServiceAddresses(
+func recordServicesWithMatchingAddress(
 	a kubectlplugin.HandlerArgs,
 	client gocoreV1.CoreV1Interface,
+	ipAddress string,
 	tableBuilder *kubectlplugin.TableBuilder,
 ) error {
 	services, err := client.Services(a.Namespace).List(a.Cmd.Context(), metaV1.ListOptions{})
 	if err != nil {
 		return err
 	}
-
 	for _, service := range services.Items {
-		for _, ip := range service.Spec.ClusterIPs {
-			if ip == "" || ip == "None" {
-				continue
-			}
-			tableBuilder.AddRow(&service, map[string]any{
-				"Kind":    "Service",
-				"Type":    "ClusterIP",
-				"Address": ip,
-			})
+		if slices.Contains(service.Spec.ClusterIPs, ipAddress) {
+			tableBuilder.AddRow(&service, map[string]any{"Type": "ClusterIP", "Kind": "Service"})
 		}
-
-		for _, ip := range service.Spec.ExternalIPs {
-			tableBuilder.AddRow(&service, map[string]any{
-				"Kind":    "Service",
-				"Type":    "ExternalIP",
-				"Address": ip,
-			})
+		if slices.Contains(service.Spec.ExternalIPs, ipAddress) {
+			tableBuilder.AddRow(&service, map[string]any{"Type": "ExternalIP", "Kind": "Service"})
 		}
-
-		if service.Spec.LoadBalancerIP != "" {
-			tableBuilder.AddRow(&service, map[string]any{
-				"Kind":    "Service",
-				"Type":    "LoadBalancerIP",
-				"Address": service.Spec.LoadBalancerIP,
-			})
+		if service.Spec.LoadBalancerIP == ipAddress {
+			tableBuilder.AddRow(&service, map[string]any{"Type": "LoadBalancerIP", "Kind": "Service"})
 		}
 	}
 	return nil
 }
 
-func getPodAddresses(
+func recordPodsWithMatchingAddress(
 	a kubectlplugin.HandlerArgs,
 	client gocoreV1.CoreV1Interface,
+	ipAddress string,
 	tableBuilder *kubectlplugin.TableBuilder,
 ) error {
 	pods, err := client.Pods(a.Namespace).List(a.Cmd.Context(), metaV1.ListOptions{})
@@ -115,20 +109,19 @@ func getPodAddresses(
 		return err
 	}
 	for _, pod := range pods.Items {
-		for _, ip := range pod.Status.PodIPs {
-			tableBuilder.AddRow(&pod, map[string]any{
-				"Kind":    "Pod",
-				"Type":    "PodIP",
-				"Address": ip.IP,
-			})
+		if slices.ContainsFunc(pod.Status.PodIPs, func(p coreV1.PodIP) bool {
+			return p.IP == ipAddress
+		}) {
+			tableBuilder.AddRow(&pod, map[string]any{"Type": "PodIP", "Kind": "Pod"})
 		}
 	}
 	return nil
 }
 
-func getNodeAddresses(
+func recordNodesWithMatchingAddress(
 	a kubectlplugin.HandlerArgs,
 	client gocoreV1.CoreV1Interface,
+	ipAddress string,
 	tableBuilder *kubectlplugin.TableBuilder,
 ) error {
 	nodes, err := client.Nodes().List(a.Cmd.Context(), metaV1.ListOptions{})
@@ -136,15 +129,10 @@ func getNodeAddresses(
 		return err
 	}
 	for _, node := range nodes.Items {
-		for _, address := range node.Status.Addresses {
-			if address.Type != coreV1.NodeInternalIP && address.Type != coreV1.NodeExternalIP {
-				continue
+		for _, n := range node.Status.Addresses {
+			if n.Address == ipAddress {
+				tableBuilder.AddRow(&node, map[string]any{"Type": n.Type, "Kind": "Node"})
 			}
-			tableBuilder.AddRow(&node, map[string]any{
-				"Kind":    "Node",
-				"Type":    address.Type,
-				"Address": address.Address,
-			})
 		}
 	}
 	return nil
