+++ cmd/kubectl-plugins/kubectl-lookup_address/main_test.go
@@ -1,4 +1,4 @@
-package kubectl_lookup_address
+package main
 
 import (
 	"context"
@@ -32,7 +32,7 @@ func TestNewCommand(t *testing.T) {
 			client := fake.NewClientset()
 			loadResources(t, client, "test")
 
-			cmd := NewCommand(kubectlplugin.WithKubeApi(client))
+			cmd := newCommand(kubectlplugin.WithKubeApi(client))
 			cmd.SetArgs([]string{"-n", "test", ip})
 			assert.NoError(cmd.Execute())
 		})
@@ -42,7 +42,7 @@ func TestNewCommand(t *testing.T) {
 		client := fake.NewClientset()
 		loadResources(t, client, "test")
 
-		cmd := NewCommand(kubectlplugin.WithKubeApi(client))
+		cmd := newCommand(kubectlplugin.WithKubeApi(client))
 		cmd.SetArgs([]string{})
 		assert.Error(cmd.Execute())
 	})
@@ -55,7 +55,7 @@ func TestNewCommand(t *testing.T) {
 				kubetesting.LoadCannedError(t, client, "list", resource)
 				loadResources(t, client, "test")
 
-				cmd := NewCommand(kubectlplugin.WithKubeApi(client))
+				cmd := newCommand(kubectlplugin.WithKubeApi(client))
 				cmd.SetArgs([]string{"-n", "test", "1.1.1.1"})
 				assert.NoError(cmd.Execute())
 			})
