+++ cmd/kubectl-plugins/kubectl-trigger_cronjob/main.go
@@ -17,10 +17,11 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package kubectl_trigger_cronjob
+package main
 
 import (
 	"fmt"
+	"os"
 
 	"github.com/ginolatorilla/devops/pkg/kubectlplugin"
 	"github.com/spf13/cobra"
@@ -29,7 +30,15 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 )
 
-func NewCommand(runnerOpts ...kubectlplugin.RunnerOpts) *cobra.Command {
+func main() {
+	if err := newCommand(kubectlplugin.WithDefaultKubeApi()).Execute(); err != nil {
+		fmt.Fprintf(os.Stderr, "error: %v\n", err)
+		os.Exit(1)
+	}
+}
+
+func newCommand(runnerOpts ...kubectlplugin.RunnerOpts) *cobra.Command {
+
 	return kubectlplugin.
 		NewRunner(triggerCronJob, append(runnerOpts, kubectlplugin.WithResourcePrinters())...).
 		ToCobraCommand(
