+++ cmd/kubectl-plugins/kubectl-trigger_cronjob/main_test.go
@@ -1,4 +1,4 @@
-package kubectl_trigger_cronjob
+package main
 
 import (
 	"context"
@@ -21,7 +21,7 @@ func TestNewCommand(t *testing.T) {
 		client := fake.NewClientset()
 		loadResources(t, client, "test")
 
-		cmd := NewCommand(kubectlplugin.WithKubeApi(client))
+		cmd := newCommand(kubectlplugin.WithKubeApi(client))
 		cmd.SetArgs([]string{"-n", "test", "test-cronjob"})
 		assert.NoError(cmd.Execute())
 	})
@@ -30,7 +30,7 @@ func TestNewCommand(t *testing.T) {
 		client := fake.NewClientset()
 		loadResources(t, client, "test")
 
-		cmd := NewCommand(kubectlplugin.WithKubeApi(client))
+		cmd := newCommand(kubectlplugin.WithKubeApi(client))
 		cmd.SetArgs([]string{"-n", "test", "missing-cronjob"})
 		assert.Error(cmd.Execute())
 	})
@@ -40,7 +40,7 @@ func TestNewCommand(t *testing.T) {
 		kubetesting.LoadCannedError(t, client, "create", "jobs")
 		loadResources(t, client, "test")
 
-		cmd := NewCommand(kubectlplugin.WithKubeApi(client))
+		cmd := newCommand(kubectlplugin.WithKubeApi(client))
 		cmd.SetArgs([]string{"-n", "test", "test-cronjob"})
 		assert.Error(cmd.Execute())
 	})
