--- cmd/testdata/tls.crt
@@ -1,1 +1,0 @@
-../kubectl_list_certs/testdata/ok/tls.crt
